fix canary build + clean env in Docker
