+++ docs/docs/dev_docs/tutorials/uniswap/typescript_glue_code.md
@@ -58,6 +58,7 @@ export FORK_URL=<YOUR_RPC_URL e.g. https://mainnet.infura.io/v3/API_KEY>
 ```
 
 Now rerun the sandbox:
+
 ```sh
 /bin/sh -c "$(curl -fsSL 'https://sandbox.aztec.network')"
 ```
@@ -94,7 +95,7 @@ import {
   parseEther,
 } from "viem";
 import { foundry } from "viem/chains";
-import { CrossChainTestHarness } from "./fixtures/cross_chain_test_harness.js";
+import { CrossChainTestHarness } from "./shared/cross_chain_test_harness.js";
 import { UniswapContract } from "./fixtures/Uniswap.js";
 import { beforeAll, expect, jest } from "@jest/globals";
 import {
@@ -113,18 +114,18 @@ const MNEMONIC = "test test test test test test test test test test test junk";
 const hdAccount = mnemonicToAccount(MNEMONIC);
 const expectedForkBlockNumber = 17514288;
 
-#include_code uniswap_l1_l2_test_setup_const yarn-project/end-to-end/src/canary/uniswap_l1_l2.ts typescript raw
-#include_code uniswap_setup yarn-project/canary/src/uniswap_trade_on_l1_from_l2.test.ts typescript raw
-#include_code uniswap_l1_l2_test_beforeAll yarn-project/end-to-end/src/canary/uniswap_l1_l2.ts typescript raw
+#include_code uniswap_l1_l2_test_setup_const yarn-project/end-to-end/src/shared/uniswap_l1_l2.ts typescript raw
+#include_code uniswap_setup yarn-project/shared/src/uniswap_trade_on_l1_from_l2.test.ts typescript raw
+#include_code uniswap_l1_l2_test_beforeAll yarn-project/end-to-end/src/shared/uniswap_l1_l2.ts typescript raw
 ```
 
 ## Private flow test
 
-#include_code uniswap_private yarn-project/end-to-end/src/canary/uniswap_l1_l2.ts typescript
+#include_code uniswap_private yarn-project/end-to-end/src/shared/uniswap_l1_l2.ts typescript
 
 ## Public flow test
 
-#include_code uniswap_public yarn-project/end-to-end/src/canary/uniswap_l1_l2.ts typescript
+#include_code uniswap_public yarn-project/end-to-end/src/shared/uniswap_l1_l2.ts typescript
 
 ## Running the test
