+++ yarn-project/canary/Dockerfile
@@ -1,16 +1,23 @@
 FROM 278380418400.dkr.ecr.eu-west-2.amazonaws.com/yarn-project AS builder
 
+COPY /usr/src/yarn-project/end-to-end /usr/src/end-to-end
+COPY /usr/src/yarn-project/end-to-end /usr/src/end-to-end
+
 # Setup & Build end-to-end package
-WORKDIR /usr/src/yarn-project/end-to-end
+WORKDIR /usr/src/end-to-end
 # Extract canary @aztec dependencies
 RUN ../canary/scripts/extract_packages.sh ../canary/package.json > ./target_pkgs.txt
 # Update end-to-end dependencies with target version
 RUN ./scripts/setup_canary.sh canary ./target_pkgs.txt
 RUN rm ./target_pkgs.txt
+# Need to yarn install & rebuild to apply changes (canary packages)
+RUN yarn && yarn build
 
-# Build canary package
-WORKDIR /usr/src/yarn-project/canary
+# Install & build canary in different dir
+COPY --from=builder /usr/src/yarn-project/canary /usr/src/canary
+WORKDIR /usr/src/canary
 RUN ./scripts/update_packages.sh canary
+RUN yarn && yarn build
 
 FROM node:18-alpine AS intermediate
 RUN apk update && apk add --no-cache udev ttf-freefont chromium curl jq bash
@@ -18,22 +25,14 @@ ENV CHROME_BIN="/usr/bin/chromium-browser" PUPPETEER_SKIP_CHROMIUM_DOWNLOAD="tru
 
 COPY --from=builder /usr/src/ /usr/src/
 WORKDIR /usr/src/yarn-project/canary
-RUN yarn && yarn build
-
-RUN apk update && apk add --no-cache udev ttf-freefont chromium curl jq bash
-ENV CHROME_BIN="/usr/bin/chromium-browser" PUPPETEER_SKIP_CHROMIUM_DOWNLOAD="true"
-
-# Copy web artifacts for browser test
-WORKDIR /usr/src/yarn-project/aztec.js
-RUN yarn build:web
 
 WORKDIR /usr/src/yarn-project/canary
 # Copy browser script
 RUN cp ../end-to-end/scripts/start_e2e_ci_browser.sh ./scripts/start_e2e_ci_browser.sh
 RUN chmod +x scripts/start_e2e_ci_browser.sh
 
-RUN cp ../aztec.js/dest/main.js src/web/
-RUN cp ../circuits.js/resources/aztec3-circuits.wasm src/web/
+RUN cp ./node_modules/@aztec/aztec.js/dest/main.js src/web/
+RUN cp ./node_modules/@aztec/circuits.js/resources/aztec3-circuits.wasm src/web/
 
 # Fresh image to copy over just the canary folder
 FROM node:18-alpine
