+++ yarn-project/canary/tsconfig.json
@@ -23,7 +23,13 @@
       "path": "../aztec.js"
     },
     {
+      "path": "../cli"
+    },
+    {
       "path": "../end-to-end"
+    },
+    {
+      "path": "../noir-contracts"
     }
   ],
   "include": ["src"]
