+++ yarn-project/end-to-end/src/e2e_cross_chain_messaging.test.ts
@@ -9,8 +9,8 @@ import {
 } from '@aztec/aztec.js';
 import { TokenBridgeContract, TokenContract } from '@aztec/noir-contracts/types';
 
-import { CrossChainTestHarness } from './fixtures/cross_chain_test_harness.js';
 import { delay, setup } from './fixtures/utils.js';
+import { CrossChainTestHarness } from './shared/cross_chain_test_harness.js';
 
 describe('e2e_cross_chain_messaging', () => {
   let logger: DebugLogger;
