+++ yarn-project/end-to-end/src/index.ts
@@ -1,5 +1,5 @@
 // Should only export tests from the canary directory
 
-export { cliTestSuite } from './canary/cli.js';
-export { browserTestSuite } from './canary/browser.js';
-export { uniswapL1L2TestSuite, UniswapSetupContext } from './canary/uniswap_l1_l2.js';
+export { cliTestSuite } from './shared/cli.js';
+export { browserTestSuite } from './shared/browser.js';
+export { uniswapL1L2TestSuite, UniswapSetupContext } from './shared/uniswap_l1_l2.js';
