+++ yarn-project/end-to-end/src/shared/browser.ts
