+++ yarn-project/end-to-end/src/uniswap_trade_on_l1_from_l2.test.ts
@@ -1,5 +1,5 @@
-import { UniswapSetupContext, uniswapL1L2TestSuite } from './canary/uniswap_l1_l2.js';
 import { setup as e2eSetup } from './fixtures/utils.js';
+import { UniswapSetupContext, uniswapL1L2TestSuite } from './shared/uniswap_l1_l2.js';
 
 // This tests works on forked mainnet. There is a dump of the data in `dumpedState` such that we
 // don't need to burn through RPC requests.
